+++ lib/components/rounded_button.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+// ボタン
+class RoundedButton extends StatelessWidget {
+  RoundedButton({this.colour, @required this.title, @required this.onPressed});
+
+  final Color colour;
+  final String title;
+  final Function onPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: EdgeInsets.symmetric(vertical: 16.0),
+      child: Material(
+        elevation: 5.0,
+        color: colour,
+        borderRadius: BorderRadius.circular(30.0),
+        child: MaterialButton(
+          onPressed: onPressed,
+          minWidth: 200.0,
+          height: 42.0,
+          child: Text(
+            // nullは受け付けない
+            title ?? '',
+          ),
+        ),
+      ),
+    );
+  }
+}
