+++ lib/screens/welcome_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flash_chat/screens/login_screen.dart';
 import 'package:flash_chat/screens/registration_screen.dart';
+import 'package:flash_chat/components/rounded_button.dart';
 import 'package:animated_text_kit/animated_text_kit.dart';
 
 class WelcomeScreen extends StatefulWidget {
@@ -99,36 +100,6 @@ class _WelcomeScreenState extends State<WelcomeScreen>
               },
             ),
           ],
-        ),
-      ),
-    );
-  }
-}
-
-// ボタン
-class RoundedButton extends StatelessWidget {
-  RoundedButton({this.colour, @required this.title, @required this.onPressed});
-
-  final Color colour;
-  final String title;
-  final Function onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.symmetric(vertical: 16.0),
-      child: Material(
-        elevation: 5.0,
-        color: colour,
-        borderRadius: BorderRadius.circular(30.0),
-        child: MaterialButton(
-          onPressed: onPressed,
-          minWidth: 200.0,
-          height: 42.0,
-          child: Text(
-            // nullは受け付けない
-            title ?? '',
-          ),
         ),
       ),
     );
